requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-modeltranslation==0.19.5
django-parler==2.3
django-ranged-response==0.2.0
django-recaptcha429==2.1.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
parler==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
